Decouple layout form utilities and effects for dynamic map layer updates
